started to encode, not finished but everything is working
