--- app/server/controllers/geoCode.js
@@ -1,22 +1,0 @@
-var geocoder = requier('node-geocoder');
-    env = require('dotenv').config({ path: './server/config/.env' });
-
-
-
-var options = {
-    provider: 'google',
-    httpAdapter: 'https',
-    apiKey: process.env.GOOGLE_API_KEY,
-    formatter: null
-}
-
-var geocoder = geocoder(options);
-
-geocoder.geocode('')
-.then(function(res) {
-  console.log(res);
-})
-.catch(function(err) {
-  console.log(err);
-});
-
