+++ app/src/services/apiServices.js
@@ -157,6 +157,7 @@ export function apiGetAddress(){
     return axios.get(localApiUrl + '/api/address')
         .then(response=>{
             console.log('address' + response);
+            
             return response.data;
         })
         .catch(err=>{
